The sweep: an unbalanced transaction could pass as balanced
⛔⛔ THE KERNEL'S SOLE INVARIANT WAS DEFEATED BY ADDITION. `transaction_total`
summed postings with `.sum()` on `i64`:

    postings   [i64::MAX, i64::MAX, 2]
    true sum   18446744073709551616
    i64 .sum() 0            ← and `transaction_is_balanced` returned TRUE

Measured. In a debug build `.sum()` panics; in release it WRAPS, and an entry
that plainly does not balance passes the door of a ledger whose entire claim is
that it cannot go out of balance. `Ratio.Core.Balanced` could not see it: in
`Int` the sum is 18446744073709551616 and is not zero.

⭐ AND BALANCE IS DECIDABLE WITHOUT OVERFLOW, which is why the fix returns a
number rather than an `Option`. The sum of any realistic count of `i64`s is
exact in `i128` — it would take 2^64 postings to trouble it — so the door check
stays INFALLIBLE and becomes correct. A guard that could error would be a door
that fails closed on a valid entry, which is its own kind of wrong. The extremes
test proves it: MAX + MIN + 1 still balances.

`transaction_total` is now `Option<i64>`, separate from the exact `i128`,
because a caller that wants a number to POST needs it to fit and a caller asking
whether an entry balances does not. Collapsing those is how the balance check
came to depend on the total being representable.

⛔ AND THE SAME SHAPE IN EVERY FOLD. `fold_nav` and `Projection::Totals`
accumulate over the WHOLE journal — `debits` adds the magnitude of every posting
ever made, so it grows with HISTORY rather than with the fund and is the first
to go. Both now accumulate in `i128` and refuse to REPORT a figure that will not
fit, rather than truncating one. A wrapped total does not look wrong; it looks
like a NAV.

⚠ And `-p.amount` was never the magnitude: `-i64::MIN` overflows. Widened
first, it does not — a second reason this is not merely about headroom.

Also fixed earlier in the sweep: `per_share(x, 0)` and `per_share(MIN, -1)`,
both live panics on the NAV path.

+++ crates/ratio-store/src/lib.rs
@@ -538,7 +538,7 @@ impl Journal for FileBook {
         // so there is no state in which the book is out and somebody has to
         // find it.
         if !entry.is_balanced() {
-            let net: i64 = entry.postings.iter().map(|p| p.amount).sum();
+            let net: i128 = entry.postings.iter().map(|p| p.amount as i128).sum();
             return Err(anyhow!(
                 "entry {:?} does not conserve value: postings net to {net}, not 0",
                 entry.id
@@ -589,7 +589,11 @@ impl Journal for FileBook {
         let mut checked: std::collections::BTreeSet<&Digest> = Default::default();
         for entry in entries {
             if !entry.is_balanced() {
-                let net: i64 = entry.postings.iter().map(|p| p.amount).sum();
+                // ⛔ `i128`. Reporting the net of an entry that did not balance
+                // must not itself overflow — the whole reason it did not balance
+                // may be that the figures are enormous, which is exactly when an
+                // `i64` sum wraps and the message says "nets to 0".
+                let net: i128 = entry.postings.iter().map(|p| p.amount as i128).sum();
                 return Err(anyhow!(
                     "entry {:?} does not conserve value: postings net to {net}, not 0",
                     entry.id
